+++ packages/swapkit/core/src/index.ts
@@ -2,3 +2,4 @@ export * from "@swapkit/api";
 export * from "@swapkit/helpers";
 
 export * from "./client.ts";
+export * from "./types.ts";