+++ packages/swapkit/helpers/src/types/derivationPath.ts
@@ -27,6 +27,7 @@ export enum DerivationPath {
   MATIC = "m/44'/60'/0'/0",
   MAYA = "m/44'/931'/0'/0",
   OP = "m/44'/60'/0'/0",
+  XRD = "////",
   THOR = "m/44'/931'/0'/0",
 }
 
@@ -51,6 +52,7 @@ export const NetworkDerivationPath: Record<Chain, DerivationPathArray> = {
   THOR: [44, 931, 0, 0, 0],
 
   // Polkadot and related network derivation path is not number based
+  XRD: [0, 0, 0, 0, 0],
   DOT: [0, 0, 0, 0, 0],
   FLIP: [0, 0, 0, 0, 0],
 };