+++ packages/wallets/wc/src/walletconnect.ts
@@ -259,7 +259,14 @@ async function getWalletconnect(
       (namespace: Todo) => namespace.accounts,
     );
 
-    return { session, accounts, client };
+    const disconnect = async () => {
+      await client.disconnect({
+        topic: session.topic,
+        reason: { code: 0, message: "User disconnected" },
+      });
+    };
+
+    return { session, accounts, client, disconnect };
   } catch (e) {
     console.error(e);
   } finally {
@@ -329,6 +336,7 @@ function connectWalletconnect({
 
       addChain({
         ...toolbox,
+        disconnect: walletconnect.disconnect,
         address,
         balance: [],
         chain,