+++ packages/toolboxes/cosmos/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @swapkit/toolbox-cosmos
 
+## 1.0.0-rc.136
+
+### Patch Changes
+
+- Updated dependencies [[`dd7c87e`](https://github.com/thorswap/SwapKit/commit/dd7c87e457fcc275d34c5aee0a08ebf3a9d47612)]:
+  - @swapkit/helpers@1.0.0-rc.110
+  - @swapkit/api@1.0.0-rc.78
+
+## 1.0.0-rc.135
+
+### Patch Changes
+
+- Updated dependencies [[`aa10e7a`](https://github.com/thorswap/SwapKit/commit/aa10e7a847c18abacbbdbefec8a93487cff2a524)]:
+  - @swapkit/helpers@1.0.0-rc.109
+  - @swapkit/api@1.0.0-rc.77
+
+## 1.0.0-rc.134
+
+### Patch Changes
+
+- Updated dependencies [[`f860028`](https://github.com/thorswap/SwapKit/commit/f860028cf61f4818d74b57345ae643d6af07007e)]:
+  - @swapkit/helpers@1.0.0-rc.108
+  - @swapkit/api@1.0.0-rc.76
+
+## 1.0.0-rc.133
+
+### Patch Changes
+
+- Updated dependencies [[`e72a9b3`](https://github.com/thorswap/SwapKit/commit/e72a9b3d8e6874175a3c1871a0b7528fb249cd18)]:
+  - @swapkit/api@1.0.0-rc.75
+  - @swapkit/helpers@1.0.0-rc.107
+
+## 1.0.0-rc.132
+
+### Patch Changes
+
+- Updated dependencies [[`f389990`](https://github.com/thorswap/SwapKit/commit/f389990a4a14566315fe353a464d916fc247ce89)]:
+  - @swapkit/helpers@1.0.0-rc.106
+  - @swapkit/api@1.0.0-rc.74
+
+## 1.0.0-rc.131
+
+### Patch Changes
+
+- Updated dependencies [[`aa93dc5`](https://github.com/thorswap/SwapKit/commit/aa93dc509e6b8e0c3cf2615d072ab7f059784c5c)]:
+  - @swapkit/helpers@1.0.0-rc.105
+  - @swapkit/api@1.0.0-rc.73
+
 ## 1.0.0-rc.130
 
 ### Patch Changes