+++ packages/toolboxes/substrate/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @swapkit/toolbox-substrate
 
+## 1.0.0-rc.45
+
+### Patch Changes
+
+- Updated dependencies [[`dd7c87e`](https://github.com/thorswap/SwapKit/commit/dd7c87e457fcc275d34c5aee0a08ebf3a9d47612)]:
+  - @swapkit/helpers@1.0.0-rc.110
+
+## 1.0.0-rc.44
+
+### Patch Changes
+
+- Updated dependencies [[`aa10e7a`](https://github.com/thorswap/SwapKit/commit/aa10e7a847c18abacbbdbefec8a93487cff2a524)]:
+  - @swapkit/helpers@1.0.0-rc.109
+
+## 1.0.0-rc.43
+
+### Patch Changes
+
+- Updated dependencies [[`f860028`](https://github.com/thorswap/SwapKit/commit/f860028cf61f4818d74b57345ae643d6af07007e)]:
+  - @swapkit/helpers@1.0.0-rc.108
+
+## 1.0.0-rc.42
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/helpers@1.0.0-rc.107
+
+## 1.0.0-rc.41
+
+### Patch Changes
+
+- Updated dependencies [[`f389990`](https://github.com/thorswap/SwapKit/commit/f389990a4a14566315fe353a464d916fc247ce89)]:
+  - @swapkit/helpers@1.0.0-rc.106
+
+## 1.0.0-rc.40
+
+### Patch Changes
+
+- Updated dependencies [[`aa93dc5`](https://github.com/thorswap/SwapKit/commit/aa93dc509e6b8e0c3cf2615d072ab7f059784c5c)]:
+  - @swapkit/helpers@1.0.0-rc.105
+
 ## 1.0.0-rc.39
 
 ### Patch Changes